// apps/enterprise-service-management/src/generated/models/field-permission-model.ts
import type { Permission } from './permission-model';
import type { Role } from './role-model';

export const FieldPermissionStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type FieldPermissionStatusKey = keyof typeof FieldPermissionStatusKeyToLabel;

export interface FieldPermission {
  /**
   * @displayName Field Permission ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Field Permission Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  fieldPermissionName: string;
  /**
   * @displayName CanCreate
   * @validationRule Required for create/update operations
   */
  canCreate: boolean;
  /**
   * @displayName CanRead
   * @validationRule Required for create/update operations
   */
  canRead: boolean;
  /**
   * @displayName CanUpdate
   * @validationRule Required for create/update operations
   */
  canUpdate: boolean;
  /**
   * @displayName EntityName
   * @validationRule Required for create/update operations
   */
  entityName: string;
  /**
   * @displayName FieldName
   * @validationRule Required for create/update operations
   */
  fieldName: string;
  /**
   * @displayName PermissionId
   * @validationRule Required for create/update operations
   */
  permissionId: Pick<Permission, 'id' | 'permissionName'>;
  /**
   * @displayName RoleId
   * @validationRule Required for create/update operations
   */
  roleId: Pick<Role, 'id' | 'roleName'>;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: FieldPermissionStatusKey;
}

export const _FieldPermission = 'FieldPermission' as const;