// apps/enterprise-service-management/src/generated/models/service-permission-model.ts
import type { Permission } from './permission-model';
import type { Role } from './role-model';
import type { Workspace } from './workspace-model';

export const ServicePermissionStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type ServicePermissionStatusKey = keyof typeof ServicePermissionStatusKeyToLabel;

export interface ServicePermission {
  /**
   * @displayName Service Permission ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Service Permission Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  servicePermissionName: string;
  /**
   * @displayName CatalogItemCode
   */
  catalogItemCode?: string;
  /**
   * @displayName PermissionId
   * @validationRule Required for create/update operations
   */
  permissionId: Pick<Permission, 'id' | 'permissionName'>;
  /**
   * @displayName RoleId
   * @validationRule Required for create/update operations
   */
  roleId: Pick<Role, 'id' | 'roleName'>;
  /**
   * @displayName ServiceCode
   * @validationRule Required for create/update operations
   */
  serviceCode: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: ServicePermissionStatusKey;
  /**
   * @displayName WorkspaceId
   * @validationRule Required for create/update operations
   */
  workspaceId: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _ServicePermission = 'ServicePermission' as const;